requirements.txt
fastapi
uvicorn
langchain
langchain-openai
faiss-cpu
transformers
sentence-transformers
python-dotenv
openrouter
gitpython